themes as parameterized page loaders
